chore: fix clippy

+++ state-chain/chains/src/btc.rs
@@ -157,7 +157,7 @@ impl FeeEstimationApi<Bitcoin> for BitcoinTrackedData {
 		&self,
 		_asset: <Bitcoin as Chain>::ChainAsset,
 		_gas_budget: GasAmount,
-		_message_length: u128,
+		_message_length: usize,
 	) -> Option<<Bitcoin as Chain>::ChainAmount> {
 		None
 	}
