+++ state-chain/chains/src/eth.rs
@@ -81,10 +81,12 @@ impl EthereumTrackedData {
 			.saturating_add(self.priority_fee)
 	}
 
+	// TODO: Does the CCM_GAS_OVERHEAD change if more calldata is passed? So basically
+	// depending on the length of the message?
 	pub fn calculate_ccm_gas_limit(
 		&self,
 		gas_budget: GasAmount,
-		_message_length: u128,
+		_message_length: usize,
 	) -> GasAmount {
 		use crate::eth::fees::*;
 		gas_budget.saturating_add(CCM_GAS_OVERHEAD).min(MAX_GAS_LIMIT)
@@ -147,7 +149,7 @@ impl FeeEstimationApi<Ethereum> for EthereumTrackedData {
 		&self,
 		_asset: <Ethereum as Chain>::ChainAsset,
 		gas_budget: GasAmount,
-		message_length: u128,
+		message_length: usize,
 	) -> Option<<Ethereum as Chain>::ChainAmount> {
 		let gas_limit = self.calculate_ccm_gas_limit(gas_budget, message_length);
 		Some(self.calculate_transaction_fee(gas_limit))
